+++ packages/react-universal/src/hooks/useLocation.ts
@@ -5,10 +5,9 @@ import { RouteLocation } from "@curi/types";
 
 export default function useLocation(props: RouteLocation) {
   const router = useRouter();
-  const pathname = React.useMemo(() => {
-    return props.name ? router.route.pathname(props.name, props.params) : "";
-  }, [props.name, JSON.stringify(props.params)]);
-
+  const pathname = props.name
+    ? router.route.pathname(props.name, props.params)
+    : "";
   return {
     pathname,
     hash: props.hash,